requirements.txt
discord.py
aiomysql
captcha
python-dotenv
jishaku